reorganize codebase to better separate buildtime from runtime concerns
