+++ setup.cfg
@@ -29,6 +29,7 @@ setup_requires =
 	setuptools
 	wheel
 install_requires =
+    importlib-metadata>=0.12;python_version<"3.8"
     stevedore>=3.4
 test_requires =
     setuptools
@@ -50,7 +51,7 @@ exclude =
 
 [options.entry_points]
 distutils.commands =
-    plugins = plugin.setuptools:plugins
+    plugins = plux.build.setuptools:plugins
 egg_info.writers =
     # this is actually not a writer, it's a reader :-)
-    plux.json = plugin.setuptools:load_plux_entrypoints
+    plux.json = plux.build.setuptools:load_plux_entrypoints
